Split notebook-header components
